+++ R/zzz.R
@@ -13,6 +13,7 @@
 "py"
 
 .onLoad <- function(libname, pkgname) {
+  
   main <- NULL
   makeActiveBinding("py", env = asNamespace(pkgname), function() {
 
@@ -28,6 +29,30 @@
     main
 
   })
+  
+  # register a callback auto-flushing Python output as appropriate
+  sys <- NULL
+  addTaskCallback(function(...) {
+    
+    enabled <- getOption("reticulate.autoflush", default = TRUE)
+    if (!enabled)
+      return(TRUE)
+    
+    if (!is_python_initialized())
+      return(TRUE)
+    
+    sys <- sys %||% import("sys", convert = TRUE)
+    
+    if (!is.null(sys$stdout) && !is.null(sys$stdout$flush))
+      sys$stdout$flush()
+    
+    if (!is.null(sys$stderr) && !is.null(sys$stderr$flush))
+      sys$stderr$flush()
+    
+    TRUE
+    
+  })
+  
 }
 
 .onUnload <- function(libpath) {